login_service/requirements.txt
affine==2.4.0
alabaster==0.7.12
alembic==1.12.0
altgraph==0.17.4
aniso8601==9.0.1
appdirs==1.4.4
area==1.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.4.0
attrs==21.2.0
Automat==22.10.0
awscli==1.22.34
Babel==2.9.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitstring==4.0.2
black==19.10b0
bleach==4.1.0
blinker==1.6.2
boofuzz==0.4.2
boto3==1.29.0
botocore==1.32.0
build==0.10.0
buildozer==1.5.0
cachetools==5.0.0
capstone==5.0.0rc2
certifi==2023.7.22
cffi==1.16.0
cgen==2020.1
chardet==5.1.0
charset-normalizer==3.3.0
ciso8601==2.3.0
cliapp==1.20180812.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmdtest==0.32+git
codepy==2023.1
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
constantly==15.1.0
contourpy==1.1.0
coverage==6.2
cryptography==41.0.4
customtkinter==5.1.3
cycler==0.11.0
Cython==0.29.33
darkdetect==0.8.0
dbus-python==1.2.18
defusedxml==0.8.0rc2
dicttoxml==1.7.4
dill==0.3.6
dirsearch==0.4.3.post1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.6
django-cors-headers==4.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dnslib==0.9.23
dnspython==2.1.0
docutils==0.16
ecdsa==0.18.0
exceptiongroup==1.1.1
execnet==1.9.0
filelock==3.12.2
flake8==6.0.0
flake8-loopy==1.1.0
flasgger==0.9.5
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
fonttools==4.40.0
funcy==2.0
genpy==2022.1
geojson==3.0.1
gitdb==4.0.9
GitPython==3.1.30
gmpy2==2.1.5
greenlet==2.0.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
imagesize==1.3.0
importlib-metadata==4.8.2
incremental==22.10.0
iniconfig==1.1.1
intervaltree==3.1.0
islpy==2023.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
keyring==23.4.0
Kivy==2.2.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
kyubi==0.1.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
localstack==2.3.2
localstack-core==2.3.2
localstack-ext==2.3.2
loopy==2020.2.1
lxml==4.8.0
M2Crypto==0.38.0
Mako==1.2.4
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.0
mechanize==0.4.7
mistune==0.8.4
more-itertools==8.10.0
mysql-connector-python==8.0.33
myst-parser==0.16.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
nodejs==0.1.1
ntlm-auth==1.5.0
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
optional-django==0.1.0
packaging==21.3
paramiko==3.2.0
pathspec==0.9.0
pbr==6.0.0
pcapy==0.11.4
pefile==2023.2.7
pep517==0.13.0
pexpect==4.8.0
pgi==0.0.11.2
Pillow==9.5.0
pkgconfig==1.5.5
pkginfo==1.8.2
platformdirs==3.5.1
pluggy==1.0.0
plumbum==1.8.2
plux==1.5.0
ply==3.11
primefac==2.0.12
protobuf==3.20.3
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pwntools==4.10.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydot==1.4.2
pyelftools==0.29
pyfiglet==0.8.post1
pyflakes==3.0.1
Pygments==2.15.1
PyGObject==3.42.1
pyinotify==0.9.6
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyJWT==2.8.0
pymbolic==2022.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.6
pyproject_hooks==1.0.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pyspnego==0.9.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-apt==2.4.0+ubuntu1
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-libpcap==0.4.2
python-magic==0.4.24
python3-openid==3.2.0
pytools==2022.1.14
pytz==2023.3.post1
PyWavelets==1.4.1
PyX==0.16
PyYAML==6.0
rasterio==1.3.7
readme-renderer==32.0
redis==3.5.3
redlibssh2==2.1.5.post1
redssh==2.3.3
regex==2021.11.10
repyexe==0.1.3
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-raw==0.2.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.4.1
roman==3.3
ROPGadget==7.3
rpyc==5.3.1
rq==1.10.1
rsa==4.8
s3transfer==0.7.0
scapy==2.5.0
scikit-image==0.21.0
scipy==1.10.1
SecretStorage==3.3.1
semver==3.0.2
service-identity==23.1.0
setuptools-scm==6.3.2
sh==1.14.3
shapely==2.0.1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
snuggs==1.4.7
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
spark-parser==1.8.9
Sphinx==4.3.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stevedore==5.1.0
systemd-python==234
tabulate==0.9.0
tailer==0.4.1
termcolor==1.1.0
tifffile==2023.4.12
tk==0.1.0
toml==0.10.2
tomli==2.0.0
tornado==6.3.3
tpm2-pytss==2.1.0
tqdm==4.62.3
ttystatus==0.38
twine==3.7.1
Twisted==22.10.0
typed-ast==1.5.1
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uncompyle6==3.9.0
unicodecsv==0.14.1
unicorn==2.0.1.post1
up42-blockutils==0.0.4
urllib3==2.0.6
uuid==1.30
validators==0.22.0
virtualenv==20.23.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
xdis==6.0.5
XlsxWriter==3.0.2
zipp==3.6.0
zope.interface==6.0
